requirements.txt
prettytable
pandas 
mysql-connector-python
